add-search v.1

+++ src/main/java/ru/yandex/practicum/filmorate/dal/ReviewLikesRepository.java
@@ -9,12 +9,12 @@ import org.springframework.stereotype.Repository;
 public class ReviewLikesRepository {
     private static final String DELETE_QUERY_COMMON = "DELETE FROM REVIEWS_LIKES WHERE REVIEW_ID = ? AND USER_ID = ?";
     private static final String DELETE_QUERY_LD = "DELETE FROM REVIEWS_LIKES WHERE REVIEW_ID = ? AND USER_ID = ?" +
-                                                  " AND IS_LIKE = ?";
+            " AND IS_LIKE = ?";
 
     private static final String INSERT_QUERY = "INSERT INTO REVIEWS_LIKES(REVIEW_ID, USER_ID, IS_LIKE) " +
-                                                "VALUES (?, ?, ?)";
+            "VALUES (?, ?, ?)";
     private static final String SELECT_USEFUL_QUERY = "select COALESCE(SUM(CASE WHEN IS_LIKE THEN 1 ELSE -1 END), 0) " +
-                                                      "USEFUL from reviews_likes where REVIEW_ID=?";
+            "USEFUL from reviews_likes where REVIEW_ID=?";
 
     private final JdbcTemplate jdbc;
 
@@ -38,7 +38,7 @@ public class ReviewLikesRepository {
     }
 
     public Integer getUsableByReviewId(Integer reviewId) {
-        return (Integer) jdbc.queryForObject(SELECT_USEFUL_QUERY, Integer.class, reviewId);
+        return jdbc.queryForObject(SELECT_USEFUL_QUERY, Integer.class, reviewId);
     }
 
     private void commonDelete(Integer reviewId, Integer userId) {
